Refactor dispatcher and exceptions into own files
