+++ dace/codegen/targets/framecode.py
@@ -7,10 +7,11 @@ import dace
 import functools
 import re
 from dace.codegen import control_flow as cflow
+from dace.codegen import dispatcher as disp
 from dace.codegen.prettycode import CodeIOStream
 from dace.codegen.targets.common import codeblock_to_cpp, sym2cpp
 from dace.codegen.targets.cpp import unparse_interstate_edge
-from dace.codegen.targets.target import TargetCodeGenerator, TargetDispatcher
+from dace.codegen.targets.target import TargetCodeGenerator
 from dace.sdfg import SDFG, SDFGState, ScopeSubgraphView
 from dace.sdfg import nodes
 from dace.sdfg.infer_types import set_default_schedule_and_storage_types
@@ -27,7 +28,7 @@ class DaCeCodeGenerator(object):
         state machines, and uses a dispatcher to generate code for
         individual states based on the target. """
     def __init__(self, *args, **kwargs):
-        self._dispatcher = TargetDispatcher()
+        self._dispatcher = disp.TargetDispatcher()
         self._dispatcher.register_state_dispatcher(self)
         self._initcode = CodeIOStream()
         self._exitcode = CodeIOStream()
