+++ dace/codegen/targets/intel_fpga.py
@@ -13,9 +13,11 @@ import dace
 from dace import registry, subsets, dtypes
 from dace.codegen import cppunparse
 from dace.config import Config
+from dace.codegen import exceptions as cgx
 from dace.codegen.codeobject import CodeObject
+from dace.codegen.dispatcher import DefinedType
 from dace.codegen.prettycode import CodeIOStream
-from dace.codegen.targets.target import make_absolute, DefinedType
+from dace.codegen.targets.target import make_absolute
 from dace.codegen.targets import cpp, fpga
 from dace.codegen.targets.common import codeblock_to_cpp
 from dace.codegen.tools.type_inference import infer_expr_type
@@ -108,7 +110,7 @@ class IntelFPGACodeGen(fpga.FPGACodeGen):
         elif execution_mode == "hardware":
             kernel_file_name += "_hardware.aocx\""
         else:
-            raise dace.codegen.codegen.CodegenError(
+            raise cgx.CodegenError(
                 "Unknown Intel FPGA execution mode: {}".format(execution_mode))
 
         host_code = CodeIOStream()
@@ -551,7 +553,7 @@ for (int u_{name} = 0; u_{name} < {size} - {veclen}; ++u_{name}) {{
                 # Due to restrictions on channel indexing, PE IDs must start from zero
                 # and skip index must be 1
                 if start_idx != 0 or skip_idx != 1:
-                    raise dace.codegen.codegen.CodegenError(
+                    raise cgx.CodegenError(
                         "Unrolled Map in {} should start from 0 and have skip equal to 1"
                         .format(sdfg.name))
                 for p in range(start_idx, stop_idx + 1, skip_idx):
